[report] check for symlink before rmtree when opt estimate-only is use
Check if the dir is also symlink before performing rmtree()
method so that unlink() method can be used instead.

Traceback (most recent call last):
  File "./bin/sos", line 22, in <module>
    sos.execute()
  File "/tmp/sos/sos/__init__.py", line 186, in execute
    self._component.execute()
OSError: Cannot call rmtree on a symbolic link

Closes: #2727

+++ sos/report/__init__.py
@@ -1059,7 +1059,7 @@ class SoSReport(SoSComponent):
             # deletion of some dirs but deleting their content
             for f in os.listdir(self.archive.get_tmp_dir()):
                 f = os.path.join(self.archive.get_tmp_dir(), f)
-                if os.path.isdir(f):
+                if os.path.isdir(f) and not os.path.islink(f):
                     rmtree(f)
                 else:
                     os.unlink(f)
